[#40]fix: classroomId로 action 초기화 위치 변경 및 actionStatus update 반환값 수정
