+++ src/main/java/soma/edupimeta/classroom/ClassroomOpenApi.java
@@ -8,6 +8,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
+import soma.edupimeta.classroom.account.models.ActionInitializeRequest;
 import soma.edupimeta.classroom.models.CreateClassroomRequest;
 import soma.edupimeta.classroom.models.MyClassroomResponse;
 import soma.edupimeta.classroom.service.domain.Classroom;
@@ -26,4 +27,11 @@ public interface ClassroomOpenApi {
         @ApiResponse(responseCode = "200", description = "클래스룸 조회 성공", content = @Content(mediaType = "application/json")),
     })
     ResponseEntity<List<MyClassroomResponse>> getMyClassrooms(Long accountId);
+
+    @Operation(summary = "클래스룸에 포함된 계정의 진척도 상태 초기화", description = "클래스룸에 포함된 계정의 진척도 상태 진행중으로 모두 초기화")
+    @ApiResponses(value = {
+        @ApiResponse(responseCode = "200", description = "초기화에 성공했습니다.", content = @Content(mediaType = "application/json")),
+        @ApiResponse(responseCode = "400", description = "초기화하는데 실패했습니다.", content = @Content(mediaType = "application/json")),
+    })
+    ResponseEntity<Long> initialization(@RequestBody ActionInitializeRequest actionInitializeRequest);
 }
