+++ src/main/java/soma/edupimeta/classroom/account/ClassroomAccountOpenApi.java
@@ -8,8 +8,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import soma.edupimeta.classroom.account.models.ActionChangeRequest;
-import soma.edupimeta.classroom.account.models.ActionInitializeRequest;
 import soma.edupimeta.classroom.account.models.CreateClassroomAccountRequest;
+import soma.edupimeta.classroom.account.service.domain.ActionStatus;
 import soma.edupimeta.classroom.account.service.domain.ClassroomAccount;
 
 @Tag(name = "ClassroomAccount", description = "ClassroomAccount API")
@@ -24,18 +24,12 @@ public interface ClassroomAccountOpenApi {
         @RequestBody CreateClassroomAccountRequest createClassroomAccountRequest
     );
 
-    @Operation(summary = "클래스룸에 포함된 계정의 진척도 상태 초기화", description = "클래스룸에 포함된 계정의 진척도 상태 진행중으로 모두 초기화")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "초기화에 성공했습니다.", content = @Content(mediaType = "application/json")),
-        @ApiResponse(responseCode = "400", description = "초기화하는데 실패했습니다.", content = @Content(mediaType = "application/json")),
-    })
-    ResponseEntity<Long> initialization(@RequestBody ActionInitializeRequest actionInitializeRequest);
-
     @Operation(summary = "클래스룸 계정의 진척도 상태 변경", description = "클래스룸 계정의 진척도 상태를 완료, 도움으로 변환")
     @ApiResponses(value = {
         @ApiResponse(responseCode = "200", description = "상태 변환에 성공했습니다.", content = @Content(mediaType = "application/json")),
         @ApiResponse(responseCode = "400", description = "상태를 변경할 수 없습니다.", content = @Content(mediaType = "application/json")),
     })
-    ResponseEntity<Void> changeClassroomAccountActionStatus(@RequestBody ActionChangeRequest actionChangeRequest);
+    ResponseEntity<ActionStatus> changeClassroomAccountActionStatus(
+        @RequestBody ActionChangeRequest actionChangeRequest);
 
 }
